requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
edge-tts==6.1.12
openai-whisper==20231117
python-multipart==0.0.12
aiofiles==24.1.0
